+++ src/index.less
@@ -621,6 +621,10 @@ label,
   }
 }
 
+.ant-pagination-item.ant-pagination-item.ant-pagination-item {
+  background-color: transparent;
+}
+
 input::-webkit-outer-spin-button,
 input::-webkit-inner-spin-button {
   -webkit-appearance: none;
